use standard file tree for images
